Capture phone number (#20)
Ignoring the name of the branch, this changes all the data captured (to include phone number, first name and last name) and fixes some bugs with the sign-up process. (Also right now some fields are optional and we are not screaming at it, I'll leave those details for later)

It also totally changes the way the SignUpView is formatted, which will transfer to LoginView later (I know it's non important but the hard-codded layout was bugging me). Experience should be pretty smooth.

WE DO NOT CONFIRM PHONE NUMBER with this PR. It will cost money and we need to figure out with Andre who's paying for that (it's cents but it would be nice to figure that out now). https://aws.amazon.com/sns/sms-pricing/


* Changed SignUpView to make users insert phone number (right now not optional).

Phone number gets captured and saved in Cognito and DynamoDB.
Also attaches the +1 country code in case the phone number doesn't have it (we're assuming only US numbers right now).

* Finished layout of SignUpView. Elements are now hidden when filling the fields